2. SAM Tool/requirements.txt
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.20.0
pandas>=1.3.0
torch>=1.10.0
torchvision>=0.11.0
pyinstaller>=4.5
cx-Freeze>=6.8
tqdm>=4.60.0
requests>=2.25.0
pytest>=6.0.0
black>=21.0.0
flake8>=3.8.0


# SAM нужен компонент (скачаем напрямую)
# git+https://github.com/facebookresearch/segment-anything.git

# МОДЕЛИ
# Самая точная
# wget https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth
# Легкая и быстрая
# wget https://dl.fbaipublicfiles.com/segment_anything/sam_vit_b_01ec64.pth